Using double precision numbers to do fee calcs

+++ src/main/scala/tech/cryptonomic/conseil/util/MathUtil.scala
@@ -6,18 +6,18 @@ object MathUtil {
 
   /**
     * Average value of a sequence of integers.
-    * @param l Sequence of integers.
+    * @param l Sequence of doubles.
     * @return
     */
-  def mean(l: Seq[Int]): Double =
+  def mean(l: Seq[Double]): Double =
     l.sum*1.0 / l.length
 
   /**
     * Standard deviation of a sequence of integers.
-    * @param l Sequence of integers/
+    * @param l Sequence of doubles.
     * @return
     */
-  def stdev(l: Seq[Int]): Double = {
+  def stdev(l: Seq[Double]): Double = {
     val m = mean(l)
     val len = l.length*1.0
     sqrt(l.map(x => (x - m)*(x - m)).sum / len)
